Run tests against MySQL too
Since this code interacts with the database, it's best to actually run
the tests against the databases that I intend to support. For now, this
adds MySQL, as well as the various mechanisms for doing the necessary
setup and running the tests against different databases.

I tried to remove as much cleverness as possible and rely on separate
rake tasks and environment variables to do the switching, as opposed to
RSpec contexts, connection pools, etc.

This seems to work well, and it creates a standard path that involves
testing against the in-memory sqlite database as the default, which
should make it a little easier to get started and run the test suite
with relatively few requirements.