Remove over-eager check for column existence
While there's value in a lot of scenarios to providing a nice error
message when you reference a column that doesn't exist, there's a bit
gotcha.

When one person adds a boolean_timestamp for a newly added column, it'll
work great for the initial development. But when another developer pulls
down that code and tries to run the migrations, they'll see this
Argument error on app boot because they haven't added the column on
their machine yet.

The simplest way around it is to remove the error checking altogether.
This will result in more opaque errors when you choose a missing column,
but I'll look at handling this in the method calls later.