Extract test database setup into a separate class
Setup was becoming increasingly complicated and more difficult to trace
and reason about when it was all dumped in spec_helper. As a separate
class, we can hide the ugliness and take advantage of statefulness to
consolidate logic, highlighting where we're actually doing something
different for a different ActiveRecord version or database engine.