Add strict option
Until now, all of the reader methods and scopes treated a
boolean_timestamp column with a future timestamp as false. While that
makes the most semantic sense, that level of precision isn't always
necessary.

Sometimes the value of the timestamp is very important. Other times,
it's a nice-to-have, a bit of data for debugging or reporting when the
column is otherwise almost entirely used like a pure boolean, leaning on
the gem to handle the time component.

In the mostly-a-boolean case, you're not usually making use of the
future date values feature (e.g. set a post to be published at a certain
time). And since querying by a specific time is generally (always?)
slower than querying for NULL or NOT NULL, you could be paying a
performance penalty for a feature you don't use.

This change adds the ability to set a boolean_timestamp as strict:
false, which changes the scopes and readers to treat future values the
same as current and past values. The only time a value is considered
false is if it's explicitly nil. By default, strict is set to true,
making this a backwards-compatible change.