Add standard for formatting
This project is tiny, so it's pretty easy to keep style fairly
consistent, but with the existence of a few tools to automatically
format our code, the hope is that I and other contributors don't have to
think about style. We can write the code and let standard do the
formatting.

I chose standard mostly because it removes any discussion about what
style rules to choose. I really don't care what the rules are, and using
standard, which isn't configurable, removes the need for those decisions
and discussions. The rubocop config it comes with is sensible enough,
and the consistency will be a win.