Lock sqlite3 to v1
ActiveRecord specifies a gem version when the adapter is loaded and the
most recent versions of Rails lock to ~> 1.4, which won't work with the
most recent sqlite3 version, which is now 2.x.

This is just for tests, so it'll be up to each app to make sure they're
using the correct sqlite version for their Rails version (as it always
has been).